// src/lib/gql/type-defs.graphqls.d.ts
export declare type RequireFields<T, K extends keyof T> = {
    [X in Exclude<keyof T, K>]?: T[X];
} & {
    [P in K]-?: NonNullable<T[P]>;
};
/** All built-in and custom scalars, mapped to their actual values */
export declare type Scalars = {
    ID: string;
    String: string;
    Boolean: boolean;
    Int: number;
    Float: number;
};
export declare type News = {
    __typename?: 'News';
    id: Scalars['ID'];
    title: Scalars['String'];
    slug: Scalars['String'];
    desc: Scalars['String'];
};
export declare type Post = {
    __typename?: 'Post';
    id: Scalars['ID'];
    title: Scalars['String'];
    slug: Scalars['String'];
    desc: Scalars['String'];
};
export declare type Query = {
    __typename?: 'Query';
    viewer: User;
    post: Post;
    news: News;
};
export declare type Mutation = {
    __typename?: 'Mutation';
    updateName: User;
    updatePost: Post;
};
export declare type MutationUpdateNameArgs = {
    name: Scalars['String'];
};
export declare type MutationUpdatePostArgs = {
    title: Scalars['String'];
};
export declare type User = {
    __typename?: 'User';
    id: Scalars['ID'];
    name: Scalars['String'];
    status: Scalars['String'];
};
export declare type ResolverTypeWrapper<T> = Promise<T> | T;
export declare type LegacyStitchingResolver<TResult, TParent, TContext, TArgs> = {
    fragment: string;
    resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export declare type NewStitchingResolver<TResult, TParent, TContext, TArgs> = {
    selectionSet: string;
    resolve: ResolverFn<TResult, TParent, TContext, TArgs>;
};
export declare type StitchingResolver<TResult, TParent, TContext, TArgs> = LegacyStitchingResolver<TResult, TParent, TContext, TArgs> | NewStitchingResolver<TResult, TParent, TContext, TArgs>;
export declare type Resolver<TResult, TParent = {}, TContext = {}, TArgs = {}> = ResolverFn<TResult, TParent, TContext, TArgs> | StitchingResolver<TResult, TParent, TContext, TArgs>;
export declare type ResolverFn<TResult, TParent, TContext, TArgs> = (parent: TParent, args: TArgs, context: TContext, info: GraphQLResolveInfo) => Promise<TResult> | TResult;
// ...
export declare type ResolversTypes = {
    News: ResolverTypeWrapper<News>;
    ID: ResolverTypeWrapper<Scalars['ID']>;
    String: ResolverTypeWrapper<Scalars['String']>;
    Post: ResolverTypeWrapper<Post>;
    Query: ResolverTypeWrapper<{}>;
    Mutation: ResolverTypeWrapper<{}>;
    User: ResolverTypeWrapper<User>;
    Boolean: ResolverTypeWrapper<Scalars['Boolean']>;
};
/** Mapping between all available schema types and the resolvers parents */
export declare type ResolversParentTypes = {
    News: News;
    ID: Scalars['ID'];
    String: Scalars['String'];
    Post: Post;
    Query: {};
    Mutation: {};
    User: User;
    Boolean: Scalars['Boolean'];
};
// ...
export declare type QueryResolvers<ContextType = any, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
    viewer?: Resolver<ResolversTypes['User'], ParentType, ContextType>;
    post?: Resolver<ResolversTypes['Post'], ParentType, ContextType>;
    news?: Resolver<ResolversTypes['News'], ParentType, ContextType>;
};
export declare type MutationResolvers<ContextType = any, ParentType extends ResolversParentTypes['Mutation'] = ResolversParentTypes['Mutation']> = {
    updateName?: Resolver<ResolversTypes['User'], ParentType, ContextType, RequireFields<MutationUpdateNameArgs, 'name'>>;
    updatePost?: Resolver<ResolversTypes['Post'], ParentType, ContextType, RequireFields<MutationUpdatePostArgs, 'title'>>;
};